Move accelerator-specific parsing functions with their accelerators
