+++ src/pytorch_lightning/accelerators/cpu.py
@@ -15,7 +15,7 @@ from typing import Any, Dict, List, Union
 
 import torch
 
-from lightning_lite.utilities.device_parser import parse_cpu_cores
+from lightning_lite.accelerators.cpu import parse_cpu_cores
 from lightning_lite.utilities.types import _DEVICE
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
