+++ src/pytorch_lightning/accelerators/tpu.py
@@ -15,7 +15,7 @@ from typing import Any, Dict, List, Optional, Union
 
 import torch
 
-from lightning_lite.utilities import device_parser
+from lightning_lite.accelerators.tpu import parse_tpu_cores
 from lightning_lite.utilities.types import _DEVICE
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.utilities.imports import _TPU_AVAILABLE
@@ -53,7 +53,7 @@ class TPUAccelerator(Accelerator):
     @staticmethod
     def parse_devices(devices: Union[int, str, List[int]]) -> Optional[Union[int, List[int]]]:
         """Accelerator device parsing logic."""
-        return device_parser.parse_tpu_cores(devices)
+        return parse_tpu_cores(devices)
 
     @staticmethod
     def get_parallel_devices(devices: Union[int, List[int]]) -> List[int]:
